requirements.txt
Django==3.2.13
asgiref==3.4.1
autopep8==1.6.0
dj-database-url==0.5.0
gunicorn==20.1.0
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
whitenoise==5.3.0
django-environ==0.8.1
python-dotenv==0.20.0

databricks-cli 
azure-storage-blob 
azure-datalake-store 
azure-mgmt-resource 
azure-mgmt-datalake-store 
azure-identity
pandas
numpy
matplotlib
pyarrow
seaborn

joblib
pathlib
bokeh
holoviews
hvplot
ennemi
statsmodels
markupsafe==2.0.1
dask[dataframe]
fmldk
celery==4.4.0
